Fix Issue499Test
Issue499Test is never run - it uses org.junit.Test, but the class is
not public, so it's never picked up.

This patch converts the class to use org.junit.jupiter.api.Test
instead so the test runner can actually pick it up, and replaces the
usage of the "expected" argument with an explicit call to
org.junit.jupirer.api.Assertions#assertThrows.
With that assertion is place it's obvious that a lot of the test's
code would never be called had the test been run, so it was safely
removed. This also helps focus the test and make it easier to
maintain.